Add more encoding tests (some failing)
