+++ cranelift/isle/veri/veri_engine/src/solver/encoded_ops/rev.rs
@@ -1,5 +1,5 @@
 use crate::solver::SolverCtx;
-use easy_smt::{Response, SExpr};
+use easy_smt::SExpr;
 
 pub fn rbit32(solver: &mut SolverCtx, x: SExpr, id: u32) -> SExpr {
     // Generated code.
@@ -530,190 +530,4 @@ pub fn rev1(solver: &mut SolverCtx, x: SExpr, id: u32) -> SExpr {
 
     let padding = solver.new_fresh_bits(solver.bitwidth - 1);
     solver.smt.concat(padding, rev1ret)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::solver::SolverCtx;
-    use std::collections::HashMap;
-    use veri_ir::TypeContext;
-
-    fn get_ctx() -> SolverCtx {
-        let smt = easy_smt::ContextBuilder::new()
-            .solver("z3", ["-smt2", "-in"])
-            .build()
-            .unwrap();
-        SolverCtx {
-            smt: smt,
-            tyctx: TypeContext {
-                tyvars: HashMap::new(),
-                tymap: HashMap::new(),
-                tyvals: HashMap::new(),
-                bv_unknown_width_sets: HashMap::new(),
-            },
-            bitwidth: 64,
-            var_map: HashMap::new(),
-            width_vars: HashMap::new(),
-            width_assumptions: vec![],
-            additional_decls: vec![],
-            additional_assumptions: vec![],
-            fresh_bits_idx: 0,
-        }
-    }
-
-    fn check(ctx: &SolverCtx, expr: SExpr, expected: &str) {
-        let expr_s = format!("{}", ctx.smt.display(expr));
-        assert_eq!(expr_s, expected);
-    }
-
-    #[test]
-    fn rev1_test() {
-        let mut ctx = get_ctx();
-
-        let x = ctx.smt.atom("x");
-        let res = rev1(&mut ctx, x, 42);
-
-        check(&ctx, res, "(concat fresh0 rev1ret_42)");
-        check(&ctx, ctx.additional_decls[0].1, "(_ BitVec 1)");
-        check(
-            &ctx,
-            ctx.additional_assumptions[0],
-            "(= rev1ret_42 ((_ extract 0 0) x))",
-        );
-    }
-
-    fn check_rev_with_solver(input: &str, output: &str, width: usize) {
-        let mut ctx = get_ctx();
-
-        // Set up an input variable
-        let ty = ctx.smt.bit_vec_sort(ctx.smt.numeral(width));
-        let input_var = ctx.declare("input".to_string(), ty);
-
-        // Set the input equal to our expected input
-        ctx.additional_assumptions
-            .push(ctx.smt.eq(input_var, ctx.smt.atom(input)));
-
-        // Call the encoding function to be tested
-        let output_from_rev = match width {
-            1 => rev1(&mut ctx, input_var, 0),
-            8 => rev8(&mut ctx, input_var, 0),
-            16 => rev16(&mut ctx, input_var, 0),
-            32 => rev32(&mut ctx, input_var, 0),
-            64 => rev64(&mut ctx, input_var, 0),
-            _ => unreachable!(),
-        };
-
-        // Extract the width of bits that we care about.
-        let output_care_bits = ctx
-            .smt
-            .extract((width - 1).try_into().unwrap(), 0, output_from_rev);
-
-        // Bookkeeping: declare declarations, assert assumptions
-        for (name, ty) in &ctx.additional_decls {
-            ctx.smt.declare_const(name, *ty).unwrap();
-        }
-        ctx.smt
-            .assert(ctx.smt.and_many(ctx.additional_assumptions.clone()))
-            .unwrap();
-
-        // Check that our expected output is valid
-        ctx.smt.push().unwrap();
-        ctx.smt
-            .assert(ctx.smt.eq(output_care_bits, ctx.smt.atom(output)))
-            .unwrap();
-        if !matches!(ctx.smt.check(), Ok(Response::Sat)) {
-            // Bad! This is a bug!
-            // Pop the output assertion
-            ctx.smt.pop().unwrap();
-            // Try again
-            assert!(matches!(ctx.smt.check(), Ok(Response::Sat)));
-            // Get the value for what output is to panic with a useful message
-            let val = ctx.smt.get_value(vec![output_care_bits]).unwrap()[0].1;
-            panic!("Expected {}, got {}", output, ctx.display_hex_to_bin(val));
-        } else {
-            ctx.smt.pop().unwrap();
-        }
-
-        // Check that there is no other possible output
-        ctx.smt.push().unwrap();
-        ctx.smt
-            .assert(
-                ctx.smt
-                    .not(ctx.smt.eq(output_care_bits, ctx.smt.atom(output))),
-            )
-            .unwrap();
-        assert!(matches!(ctx.smt.check(), Ok(Response::Unsat)));
-        ctx.smt.pop().unwrap();
-    }
-
-    #[test]
-    fn test_rev1_with_solver() {
-        check_rev_with_solver("#b0", "#b0", 1);
-        check_rev_with_solver("#b1", "#b1", 1);
-    }
-
-    #[test]
-    fn test_rev8_with_solver() {
-        check_rev_with_solver("#b01010101", "#b10101010", 8);
-        check_rev_with_solver("#b11110000", "#b00001111", 8);
-        check_rev_with_solver("#b00000000", "#b00000000", 8);
-        check_rev_with_solver("#b11111111", "#b11111111", 8);
-    }
-
-    #[test]
-    fn test_rev16_with_solver() {
-        check_rev_with_solver("#b0101010101010101", "#b1010101010101010", 16);
-        check_rev_with_solver("#b1111111100000000", "#b0000000011111111", 16);
-        check_rev_with_solver("#b0000000000000000", "#b0000000000000000", 16);
-        check_rev_with_solver("#b1111111111111111", "#b1111111111111111", 16);
-    }
-
-    #[test]
-    fn test_rev32_with_solver() {
-        check_rev_with_solver(
-            "#b01010101010101010101010101010101",
-            "#b10101010101010101010101010101010",
-            32,
-        );
-        check_rev_with_solver(
-            "#b11111111111111110000000000000000",
-            "#b00000000000000001111111111111111",
-            32,
-        );
-        check_rev_with_solver(
-            "#b00000000000000000000000000000000",
-            "#b00000000000000000000000000000000",
-            32,
-        );
-        check_rev_with_solver(
-            "#b11111111111111111111111111111111",
-            "#b11111111111111111111111111111111",
-            32,
-        );
-    }
-
-    #[test]
-    fn test_rev64_with_solver() {
-        check_rev_with_solver(
-            "#b0101010101010101010101010101010101010101010101010101010101010101",
-            "#b1010101010101010101010101010101010101010101010101010101010101010",
-            64,
-        );
-        check_rev_with_solver(
-            "#b1111111111111111111111111111111100000000000000000000000000000000",
-            "#b0000000000000000000000000000000011111111111111111111111111111111",
-            64,
-        );
-        check_rev_with_solver(
-            "#b0000000000000000000000000000000000000000000000000000000000000000",
-            "#b0000000000000000000000000000000000000000000000000000000000000000",
-            64,
-        );
-        check_rev_with_solver(
-            "#b1111111111111111111111111111111111111111111111111111111111111111",
-            "#b1111111111111111111111111111111111111111111111111111111111111111",
-            64,
-        );
-    }
 }
